pytest and dependency updates
