+++ .travis.yml
@@ -6,7 +6,6 @@ python:
 install:
     - pip install pipenv
     - pipenv install
-    - pipenv shell
 
 script:
     - pytest
