+++ python-flask/modules/app/config.py
@@ -12,6 +12,7 @@ from flask_cors import CORS
 
 UPLOAD_FOLDER = '/uploads'
 
+
 class JSONEncoder(json.JSONEncoder):
     ''' extend json-encoder class'''
 
@@ -43,3 +44,4 @@ flask_bcrypt = Bcrypt(app)
 jwt = JWTManager(app)
 # use the modified encoder class to handle ObjectId & datetime object while jsonifying the response.
 app.json_encoder = JSONEncoder
+
