+++ python-flask/modules/utils/allowedFilenames.py
@@ -1,0 +1,4 @@
+def allowed_file(filename):
+    ALLOWED_EXTENSIONS = set(['csv'])
+    return '.' in filename and \
+           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
